Refactored monotonic clock.
