+++ spec/concurrent/exchanger_spec.rb
@@ -56,7 +56,7 @@ module Concurrent
       context 'with timeout' do
 
         it 'should block until timeout' do
-          duration = Concurrent.monotonic_interval do
+          duration = Concurrent::TestHelpers.monotonic_interval do
             subject.exchange(2, 0.1)
           end
           expect(duration).to be_within(0.05).of(0.1)
